arm: Construct correct addresses for multi entry mappings when hyp enabled
Multi entry mappings (large pages and super sections) are a just a hint with the paging
format when hyp is enabled. As such the entry needs to be correct as though the hint bit
is not set, meaning each of the 16 entries needs to have an independently correct
address field. This was already done in the case of remapping, but not for mapping. This
change implements this for mapping, and changes the remapping code to use a loop
formulation that will be easier to verify.