mcs: call updatetimestamp and commitTime once
This change refactors (and simplifies) the interactions with the
scheduler on the irq path. We don't call commitTime on this path,
as this call was redundant (we'll call it in the scheduler if we need
to). Move up the calls to checkBudget and updateTimestamp.

Avoid modifying the scheduler on the unlocked IPI path