manual: Variable for python interpreter
This allows the manual to be built with `PYTHON=mypython make`, to
invoke the python scripts involved in generating documentation with a
particular version of python. The scripts are compatible with both
python2 and python3, but this should simplify debugging if this ever
stops being the case. If no value of PYTHON is specified, the default
value of "python" is used.