VER-779: another refill refactor + fix
The step to check if a refill head is > MIN_BUDGET needs to
be after the call to refill_split_check when we know we are ending the
timeslice, rather than at the end of
refill_budget_check.

This refactors the call to refill_split_check from chargeBudget into
refill_budget_check and moves the code that makes sure refill_head >
MIN_BUDGET to after that call.