Make clock check a warning only instead of a hard fault
This clock value isn't set consistently on some platforms due to software
rounding, so halting here tends to happen a lot.  Instead print a warning
and continue on.

For example using a default Xilinx platform with a 33.333... MHz oscialltor
and a default bootloader the clock value will be somewhere in the range
99999000-100000000, which for most practical purposes is the same thing.

Since this register is written by software and doesn't have the precision to
fully represent common hardware oscillator frequencies, this check seems overly
strict anyway.