cmake: Add missing bitfield generation dependency
The GenHBFTarget creates both a command and a target for a bitfield header. For correct
dependencies and parallelism you need to depend both the command *and* the target.
Currently only the target is being depended upon, meaning that rebuilds do not happen
correctly. This change adds the produced file, i.e. the command, as a dependency as well.