x86: User zero as idle thread stack sentinel
The RSP of the idle thread is checked to determine whether an interrupt occured in the idle
thread or during a pre-emptible period of kernel execution. This changes the sentinel value
to be a constant 0, instead of the address of the IRQ stack. Using zero is simpler, prevents
any relation to the stack being used to save values and is simpler for verification as it
is a clear constant value.