x64: Declare virtual addresses as 48-bits instead of 49-bits
In f6d367e8dff6a9a2c0d14bd08535bdd8a6a9a963 virtual addresses were changed from 48
bits to 49 bits, but this does not seem necessary and the commit does not have enough
detail to justify the change. As such it is being changed back to 48 bits in this
commit, which is far less confusing and results on the bitfield generated functions
switching on bit 47 to determine sign extension (instead of bit 48), which matches
what the hardware description says should be done.