x86: Correctly reset kernel stack on nested interrupt
Previous code assumed that the kernel stack was size aligned and attempted to reset the stack
pointer by masking and adding. The assumption that the kernel stack is size aligned is strictly
not true, and this lead to resetting the kernel stack to 'random' memory, and corrupting it.

The solution used here is to directly pass in a new value for the stack top.