x64: Invalid ASID with single asm block
This performs a double address space switch to invalidate an ASID using a single hand
written asm block instead of setCurrentCR3. The purpose of this is to ensure that the
only requirements on the temporary address space that we switch to is that the kernel
code is still mapped in at the same location. As we ensure our old cr3 value is in
register we can ensure, unlike with calling getCurrentCR3 that no stack or other
temporary memory is required.