x86: Prevent RSB based speculation on vmexit
A guest, due to its ability to run at the same virtual addresses as the kernel, may have
left return values in the RSB that could cause uncontrolled speculation for a Spectre
like attack. In all scenarios, with or without IBRS, we need to flush the RSB to guard
against this.