cmake: Refactor file kernel compilation using CPPFile and changing source file properties
CPPFile is a more robuse verision of GenCPPCommand and allows us to drop the
`custom_command_c_flags` as cmake can directly perform compilation.

Changing the LANGUAGE of the SOURCE file from CPPFile allows us to pass a file with a .i
name to GCC, which gcc already knows to not preprocess again. This removes the need for
the custom compilation step.

Due to external tools depending on the kernel_all_pp.c output we still need to provide
this.