x86: IOPort invocation has proper structure
Invocations should have a seperate decode and invoke stage and the previous invocation
performed stateful operations during the notional 'decode' phase, and did not use a
separate function for the invoke stage. This provides the correct structure by splitting
the decode into two halves, one for the 'in' operations, and one for the 'out' operations,
and defines an invoke for each of them.

This solution has some duplication due to repeatedly decoding the invocation label to
determine the size of the port operation, but should ultimately be the simplist solution
to verify.