x64: Different perform for Map and Remap of huge pages
Previously Map and Remap of huge (1GiB) pages, which were mode specific to x64, would
ultimately call the same perform function that would update the cap and then set the
hardware mapping. Setting the cap (or re-setting as the case is) for remap is both
redundant and not how other verified platforms implement their remap invocation. For
simplicity of verification it is easier to change perform to not update the cap in
the case of remap.