riscv/arch: sbi - Trap to riscv-pk in seL4 machine mode
seL4 resets mtval when it works in machine mode. In order to be able
to trap to riscv-pk if needed, write mtval with riscv-pk's trap
vector address.

Note, on a real hardware platform, there shouldn't be calls to riscv-pk.
For example, putting characters and setting timers will never trap
to riscv-pk as hardware (e.g. UART and timer) doesn't need to be
emulated, and thus can be accesssed directly from seL4 and/or user.