spike: Refactor timer implementations to use get_time/rdtime
get_cycles in this file does not really get cycles, it reads time.

Back in 2015, time was the same as cycles in spike/pk, but this has
changed, yet the name in the port didn't.

The commit cleans up timer set/reset to use get_time defined to read
time[h] CSR (which is emulated by riscv-pk in the case of Spike).

In machine one, we need to save the previous trap address because
reading time traps to riscv-pk on Spike, so we set the trap address
to riscv-pk.