x86: All page unmap paths perform an invocation
This takes the current logic in the decode function for unmap that decides which perform
function to call and considers that logic to be part of the perform step, and places it
in a wrapping perform function. As a result the case where we do not call any detailed
perform function will still result in calling a perform step from decode, instead of doing
nothing.