riscv: Do not zero memory in Arch_createObject
It use to be be the responsibility of createObject, and by extension Arch_createObject,
to zero memory if so required. Since the riscv port was originally started the responsiblity
of zering the memory was moved to generic untyped retype handling, however this change
was never propogated to the riscv architecture. As it stands this memory zeroing is
guaranteed to zero memory that was already zeroed, which is just a performance (and proof),
issue and so we want to remove them.