Move some lockTLB magic into C.
This patch moves the 'outer' chunk of lockTLBEntry into C rather
than handwritten assembly. The outer chunk accesses a global
counter and does arithmetic. The inner chunk (lockTLBEntryCritical)
writes to the registers, must be specially aligned, and is generally
special.

The change reduces unnecessary handwritten assembly, and also avoids
a special case that was problematic for binary verification.