riscv: move virtual memory fence after address space switch
While the RISC-V ISA says to that a VMFENCE before a write to the
satp may be necessary, vm faults occur with this ordering when
running on the Rocket Chip. Placing the VMFENCE after the satp write
resolves these faults.

This is also the ordering used in the RISC-V port of Linux when
switching MMU contexts.

Change-Id: I1ec794651d080a5e7a987fa8b2062dc01daeb683