x86_64: Rework boot code to always use small pages
Previously, there was a check for whether or not huge pages are
supported by the processor before proceeding to set up the temporary
boot page table structures with the corresponding page size.

This is unnecessary as the page table structures are destroyed and
redefined again before entering user space.

The check for CONFIG_HUGE_PAGE is retained but not used for determining
the size of the boot page table structures. Instead, small pages are
always used as this is always supported by the architecture. This is a
beneficial change as it reduces the size and complexity of the untrusted
assembly boot code.

This change has the consequence of encouraging TLB misses during the
boot process but considering the size of the kernel and the lack of
running user space processes this is negligible.