AArch64: Modify kernel higher half mapping: increase from 510 to 512 GiB
The previous 510GiB mapping was chosen to make it easier to eyeball addresses
while debugging: it was trivial to determine the physical address that a virtual
address pertained to in the kernel with this mapping.

This was no longer considered to be a necessary benefit, so this is being
changed.