timer-refactor: Move kernel timers and rename
This change moves all arm platform timers to drivers/timer
and renames them to the compatibility string specified in the dts.
Now the correct timer is included at build time based on the dts,
this follows the same pattern as the serial-refactor.
Have moved the arm generic_timer and priv_timer into the timer
folder and updated cmake configs appropriately.